Begin `unsigned long`->`size_t` conversion to support large files on Windows

+++ object-file.h
@@ -147,7 +147,7 @@ struct object_info;
 int parse_loose_header(const char *hdr, struct object_info *oi);
 
 int write_object_file(struct odb_source *source,
-		      const void *buf, unsigned long len,
+		      const void *buf, size_t len,
 		      enum object_type type, struct object_id *oid,
 		      struct object_id *compat_oid_in, unsigned flags);
 
@@ -197,7 +197,7 @@ int finalize_object_file_flags(struct repository *repo,
 			       enum finalize_object_file_flags flags);
 
 void hash_object_file(const struct git_hash_algo *algo, const void *buf,
-		      unsigned long len, enum object_type type,
+		      size_t len, enum object_type type,
 		      struct object_id *oid);
 
 /* Helper to check and "touch" a file */
